+++ backend/src/main/java/hu/projekt/bolt/service/DolgozoService.java
@@ -5,9 +5,9 @@ import java.util.List;
 
 public interface DolgozoService {
     DolgozoDTO ujDolgozo(String nev, String email, String jelszo);
-    DolgozoDTO modositDolgozo(Long id, String nev, String email, String szerepkor);
-    DolgozoDTO modositJelszo(Long id, String ujJelszo);
-    void torolDolgozo(Long id);
+    DolgozoDTO modositDolgozo(int id, String nev, String email, String szerepkor);
+    DolgozoDTO modositJelszo(int id, String ujJelszo);
+    void torolDolgozo(int id);
     List<DolgozoDTO> osszesDolgozo();
-    DolgozoDTO getDolgozoById(Long id);
+    DolgozoDTO getDolgozoById(int id);
 }
